+++ BillSplitter/iOSProjectTemplate/Functions.h
@@ -42,10 +42,10 @@
 	#define transposeCGRect(r) (CGRectMake(r.origin.y, r.origin.x, r.size.height, r.size.width))
 
 	/** Create UIColor form hex code color value */
-	#define UIColorFromHex(hex) [UIColor colorWithRed:((float)((hex & 0xFF000000) >> 24))/255.0 green:((float)((hex & 0xFF0000) >> 16))/255.0 blue:((float)((hex & 0xFF00) >> 8))/255.0 alpha:((float)(hex & 0xFF))/255.0]
+	#define UIColorFromHex(hex) [UIColor colorWithRed:((CGFloat)((hex & 0xFF000000) >> 24))/255.0 green:((CGFloat)((hex & 0xFF0000) >> 16))/255.0 blue:((CGFloat)((hex & 0xFF00) >> 8))/255.0 alpha:((CGFloat)(hex & 0xFF))/255.0]
 
-	/** Generate random float in range */
-	#define randomf(rangeStart, rangeEnd) ((((float) (arc4random() % ((unsigned)RAND_MAX + 1)) / RAND_MAX) * (rangeEnd - rangeStart)) + rangeStart)
+	/** Generate random CGFloat in range */
+	#define randomf(rangeStart, rangeEnd) ((((CGFloat) (arc4random() % ((unsigned)RAND_MAX + 1)) / RAND_MAX) * (rangeEnd - rangeStart)) + rangeStart)
 
 	/** Take out all punctuation in a string */
 	#define NSStringWithoutPunctuation(s) ([[s componentsSeparatedByCharactersInSet: [NSCharacterSet punctuationCharacterSet]] componentsJoinedByString:@""])
